+++ arbor/CMakeLists.txt
@@ -53,7 +53,6 @@ set(arbor_sources
     schedule.cpp
     spike_event_io.cpp
     spike_source_cell_group.cpp
-    swcio.cpp
     s_expr.cpp
     threading/threading.cpp
     thread_private_spike_store.cpp
